requirements.txt
# Core Deep Learning
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.24.0
scipy>=1.10.0

# Diffusion Models
diffusers>=0.21.0
accelerate>=0.20.0
transformers>=4.30.0

# Trading & Market Data
ccxt>=4.0.0  # Cryptocurrency exchange integration
pandas>=2.0.0
ta-lib>=0.4.0  # Technical analysis library
yfinance>=0.2.0  # Stock market data

# Risk Management & Portfolio
quantlib>=1.30
empyrical>=0.5.5  # Performance metrics
pyfolio>=0.9.2  # Portfolio analysis

# Data Processing
scikit-learn>=1.3.0
joblib>=1.3.0

# Visualization & Monitoring
matplotlib>=3.7.0
seaborn>=0.12.0
plotly>=5.14.0
tensorboard>=2.13.0

# Database & Storage
sqlalchemy>=2.0.0
redis>=4.5.0
pymongo>=4.4.0

# API & Web
fastapi>=0.100.0
uvicorn>=0.23.0
websockets>=11.0
requests>=2.31.0

# Utilities
python-dotenv>=1.0.0
pyyaml>=6.0
tqdm>=4.65.0
loguru>=0.7.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0

